+++ client/src/components/TestSuite/Requirements/styles.tsx
@@ -23,4 +23,13 @@ export default makeStyles()((theme: Theme) => ({
   currentItem: {
     fontWeight: 600,
   },
+  markdown: {
+    p: {
+      margin: 0,
+    },
+  },
+  textButton: {
+    fontWeight: 'bolder',
+    cursor: 'pointer',
+  },
 }));