+++ client/src/components/TestSuite/TestSession.tsx
@@ -96,7 +96,7 @@ const TestSessionComponent: FC<TestSessionComponentProps> = ({
     [testSession.test_suite],
   );
   const requirementToTests = React.useMemo(
-    () => mapRequirementToIds(requirements, testSession.test_suite),
+    () => mapRequirementToIds(testSession.test_suite),
     [testSession.test_suite],
   );
   const splitLocation = useLocation().hash.replace('#', '').split('/');
@@ -275,8 +275,6 @@ const TestSessionComponent: FC<TestSessionComponentProps> = ({
           variant: 'error',
         });
       });
-
-    // TODO: map runnable id to requirements
   };
 
   const runTests = (runnableType: RunnableType, runnableId: string) => {