+++ lib/inferno/apps/cli/main.rb
@@ -6,8 +6,8 @@ require_relative 'services'
 require_relative 'suite'
 require_relative 'suites'
 require_relative 'new'
-require_relative '../../version'
 require_relative 'execute'
+require_relative '../../version'
 
 module Inferno
   module CLI
@@ -34,6 +34,9 @@ module Inferno
 
         # Loads the us core ig and evaluate the data included in the IG's example folder, with results redirected to outcome.json as an OperationOutcome
         `bundle exec inferno evaluate ./uscore.tgz --output outcome.json`
+
+        # Loads the us core ig and evaluate the data included in the IG's example folder using a custom configuration file
+        `bundle exec inferno evaluate ./uscore.tgz --config ./custom_config.yml`
       LONGDESC
       # TODO: Add options below as arguments
       option :data_path,
@@ -45,6 +48,10 @@ module Inferno
              aliases: ['-o'],
              type: :string,
              desc: 'Export evaluation result to outcome.json as an OperationOutcome'
+      option :config,
+             aliases: ['-c'],
+             type: :string,
+             desc: 'Path to a custom configuration file'
       def evaluate(ig_path)
         Evaluate.new.run(ig_path, options[:data_path], options)
       end
@@ -92,6 +99,10 @@ module Inferno
 
       desc 'suites', 'List available test suites'
       def suites
+        ENV['NO_DB'] = 'true'
+
+        require_relative '../../../inferno'
+
         Suites.new.run
       end