+++ lib/inferno/apps/cli/requirements.rb
@@ -1,3 +1,4 @@
+require_relative 'requirements_coverage_checker'
 require_relative 'requirements_exporter'
 
 module Inferno
@@ -11,6 +12,9 @@ module Inferno
       LONGDESC
       def export_csv
         ENV['NO_DB'] = 'true'
+
+        require_relative '../../../inferno'
+
         RequirementsExporter.new.run
       end
 
@@ -21,7 +25,60 @@ module Inferno
       LONGDESC
       def check
         ENV['NO_DB'] = 'true'
+
+        require_relative '../../../inferno'
+
+        Inferno::Application.start(:requirements)
+
         RequirementsExporter.new.run_check
+      end
+
+      desc 'coverage [TEST_SUITE_ID]',
+           "Check whether all of a test suite's requirements are tested. If no test suite id is provided, " \
+           'all test suites in the current test kit will be checked.'
+      long_desc <<~LONGDESC
+        Check whether all of the requirements declared by a test suite are
+        tested by the tests in the test suite
+      LONGDESC
+      def coverage(test_suite_id = nil)
+        ENV['NO_DB'] = 'true'
+
+        require_relative '../../../inferno'
+
+        Inferno::Application.start(:requirements)
+
+        if test_suite_id.present?
+          RequirementsCoverageChecker.new(test_suite_id).run
+        else
+          Inferno::Repositories::TestSuites.all.each do |test_suite|
+            if Object.const_source_location(test_suite.to_s).first.start_with? File.join(Dir.pwd, 'lib')
+              RequirementsCoverageChecker.new(test_suite.id).run
+            end
+          end
+        end
+      end
+
+      desc 'check_coverage [TEST_SUITE_ID]',
+           'Check whether the coverage CSV files are up to date'
+      long_desc <<~LONGDESC
+        Check whether the coverage CSV files are up to date
+      LONGDESC
+      def check_coverage(test_suite_id = nil)
+        ENV['NO_DB'] = 'true'
+
+        require_relative '../../../inferno'
+
+        Inferno::Application.start(:requirements)
+
+        if test_suite_id.present?
+          RequirementsCoverageChecker.new(test_suite_id).run_check
+        else
+          Inferno::Repositories::TestSuites.all.each do |test_suite|
+            if Object.const_source_location(test_suite.to_s).first.start_with?(File.join(Dir.pwd, 'lib'))
+              RequirementsCoverageChecker.new(test_suite.id).run_check
+            end
+          end
+        end
       end
     end
   end