+++ lib/inferno/apps/web/serializers/requirement.rb
@@ -8,10 +8,12 @@ module Inferno
 
         field :requirement
         field :conformance
-        field :actor
-        field :sub_requirements
+        field :actors
+        field :subrequirements
         field :conditionality
         field :url, if: :field_present?
+        field :not_tested_reason, if: :field_present?
+        field :not_tested_details, if: :field_present?
       end
     end
   end