+++ lib/inferno/dsl/must_support_assessment.rb
@@ -234,6 +234,10 @@ module Inferno
 
         def resource_populates_element?(resource, element_definition)
           path = element_definition[:path]
+
+          # handle MustSupport element under extension: Ex: extension:supporting-info.value[x]
+          resource, path = process_must_support_element_in_extension(resource, path) if path.start_with?('extension:')
+
           ms_extension_urls = must_support_extensions.select { |ex| ex[:path] == "#{path}.extension" }
             .map { |ex| ex[:url] }
 
@@ -243,6 +247,25 @@ module Inferno
 
           # Note that false.present? => false, which is why we need to add this extra check
           value_found.present? || value_found == false
+        end
+
+        def process_must_support_element_in_extension(resource, path)
+          return [resource, path] unless path.start_with?('extension:')
+
+          path_without_prefix = path.delete_prefix('extension:')
+          extension_split = path_without_prefix.split('.')
+          extension_name = extension_split.first
+          extension_path = extension_split.last
+
+          found_extension_url = must_support_extensions.find { |ex| ex[:id].include?(extension_name) }[:url]
+          ms_element_extension = resource.extension.find { |ex| ex.url == found_extension_url }
+
+          if ms_element_extension.present?
+            resource = ms_element_extension
+            path = extension_path
+          end
+
+          [resource, path]
         end
 
         def matching_without_extensions?(value, ms_extension_urls, fixed_value)