+++ lib/inferno/entities/requirement.rb
@@ -11,8 +11,9 @@ module Inferno
         :url,
         :requirement,
         :conformance,
-        :actor,
-        :sub_requirements,
+        :actors,
+        :subrequirements,
+        :subrequirements_string,
         :conditionality,
         :not_tested_reason,
         :not_tested_details
@@ -23,7 +24,13 @@ module Inferno
       def initialize(params)
         super(params, ATTRIBUTES)
 
-        self.requirement_set = id.split('@').first if requirement_set.blank? && id&.include?('@')
+        return unless requirement_set.blank? && (id&.include?('@') || id&.include?('#'))
+
+        self.requirement_set = id.split(/[@#]/).first
+      end
+
+      def subrequirements
+        @subrequirements ||= self.class.expand_requirement_ids(subrequirements_string, requirement_set)
       end
 
       # Expand a comma-delimited list of requirement id references into an Array
@@ -35,12 +42,14 @@ module Inferno
       #   used if none is included in the `requirement_id_string`
       #
       # @example
-      # expand_requirement_ids('example-ig@1,3,5-7')
-      # # => ['example-ig@1','example-ig@3','example-ig@5','example-ig@6','example-ig@7']
-      # expand_requirement_ids('example-ig')
-      # # => []
-      # expand_requirement_ids('1,3,5-7', 'example-ig')
-      # # => ['example-ig@1','example-ig@3','example-ig@5','example-ig@6','example-ig@7']
+      #   expand_requirement_ids('example-ig@1,3,5-7')
+      #   # => ['example-ig@1','example-ig@3','example-ig@5','example-ig@6','example-ig@7']
+      #   expand_requirement_ids('example-ig')
+      #   # => []
+      #   expand_requirement_ids('1,3,5-7', 'example-ig')
+      #   # => ['example-ig@1','example-ig@3','example-ig@5','example-ig@6','example-ig@7']
+      #   expand_requirement_ids('example-ig#actor1')
+      #   # => [all requirements for actor1 from example-ig]
       def self.expand_requirement_ids(requirement_id_string, default_set = nil) # rubocop:disable Metrics/CyclomaticComplexity
         return [] if requirement_id_string.blank?
 
@@ -49,10 +58,16 @@ module Inferno
           .split(',')
           .map(&:strip)
           .flat_map do |requirement_string|
-            current_set, requirement_string = requirement_string.split('@') if requirement_string.include?('@')
+            if requirement_string.include? '@'
+              current_set, requirement_string = requirement_string.split('@')
+            elsif requirement_string.include? '#'
+              current_set, actor = requirement_string.split('#')
+            end
 
             requirement_ids =
-              if requirement_string.include? '-'
+              if actor.present?
+                return Repositories::Requirements.new.requirements_for_actor(current_set, actor).map(&:id)
+              elsif requirement_string.include?('-') && !requirement_string.match?(/[^\d\-]/)
                 start_id, end_id = requirement_string.split('-')
                 if start_id.match?(/^\d+$/) && end_id.match?(/^\d+$/)
                   (start_id..end_id).to_a
@@ -65,6 +80,10 @@ module Inferno
 
             requirement_ids.map { |id| "#{current_set}@#{id}" }
           end
+      end
+
+      def actor?(actor_to_check)
+        actors.any? { |actor| actor.casecmp? actor_to_check }
       end
 
       def tested?