+++ spec/inferno/entities/requirement_spec.rb
@@ -1,6 +1,6 @@
 RSpec.describe Inferno::Entities::Requirement do
   describe '.expand_requirement_ids' do
-    it 'returns an empty array of there are no sub_requirements' do
+    it 'returns an empty array of there are no subrequirements' do
       expect(described_class.expand_requirement_ids(nil)).to eq([])
     end
 
@@ -35,9 +35,17 @@ RSpec.describe Inferno::Entities::Requirement do
       expect(described_class.expand_requirement_ids(ids.join(','))).to eq(expected_ids)
     end
 
-    it 'ignores invalid non-numeric range boundaries' do
+    it 'does not treat hyphens in non-numeric ids as ranges' do
       ids = ['criteria3@a-b', 'criteria4@2', 'example-ig']
-      expected_ids = ['criteria4@2']
+      expected_ids = ['criteria3@a-b', 'criteria4@2', 'criteria4@example-ig']
+
+      expect(described_class.expand_requirement_ids(ids.join(','))).to eq(expected_ids)
+    end
+
+    it 'adds requirements when specified by actor' do
+      ids = ['sample-criteria-proposal-5#Client']
+      expected_ids = ['sample-criteria-proposal-5@1', 'sample-criteria-proposal-5@3']
+
       expect(described_class.expand_requirement_ids(ids.join(','))).to eq(expected_ids)
     end
   end